*** empty log message ***
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@560 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/store.c
@@ -306,7 +306,7 @@ char *object;
 	/* Write address */
 #if PTRSIZ > 4
  
-	trunc_ptr = object & 0xffff;
+	trunc_ptr = (unsigned int)object & 0xffff;
 	buffer_write (&trunc_ptr, sizeof (unsigned int));
 #else
 	buffer_write(&object, sizeof(char *));
